# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Image upload, compression, transformation, and storage management API.",
        "title": "Image Pipeline API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/batches/{batchId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Poll the progress of a batch transform or revert operation.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transforms"
                ],
                "summary": "Get batch job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Batch ID",
                        "name": "batchId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/image-pipeline_internal_models.BatchJob"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/docs.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/docs.APIResponse"
                        }
                    }
                }
            }
        },
        "/image/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a single image by ID. Only the owning user can delete.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "Delete an image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "schema": {
                            "$ref": "#/definitions/docs.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/docs.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/docs.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Delete failed",
                        "schema": {
                            "$ref": "#/definitions/docs.APIResponse"
                        }
                    }
                }
            }
        },
        "/images": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of the authenticated user's images. Supports search and status filtering.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "List images",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page (max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by filename",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/docs.PaginatedImages"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Fetch failed",
                        "schema": {
                            "$ref": "#/definitions/docs.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete multiple images by IDs. Max 50 per request.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "Batch delete images",
                "parameters": [
                    {
                        "description": "Image IDs to delete",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.BatchIDsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/docs.DeleteResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "No IDs or too many",
                        "schema": {
                            "$ref": "#/definitions/docs.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Batch delete failed",
                        "schema": {
                            "$ref": "#/definitions/docs.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/images/{id}/cancel-transform": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancel a transform that is still in \"processing\" status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transforms"
                ],
                "summary": "Cancel a pending transform",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cancelled",
                        "schema": {
                            "$ref": "#/definitions/docs.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/docs.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/docs.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "docs.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "operation completed"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "image-pipeline_internal_models.BatchJob": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/image-pipeline_internal_models.BatchError"
                    }
                },
                "failed": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "imageIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "description": "\"processing\", \"completed\", \"partial\"",
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                },
                "transformations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/image-pipeline_internal_models.TransformConfig"
                    }
                },
                "type": {
                    "description": "\"transform\" or \"revert\"",
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}